Crate coupe_tools

Modules

ittapi

Wrapper around ittapi to avoid spreading #cfg flags in the codebase.

Structs

EdgeWeightDistError

MeshFormatError

Problem

Enums

EdgeWeightDistribution

MeshFormat

Traits

ToRunner

Functions

barycentres

dual

The adjacency matrix that models the dual graph of the given mesh.

parse_algorithm

set_edge_weights

used_element_count

The number of elements that are taken into account for partitioning.

write_mesh

Write a mesh to stdout in the given format.

Type Definitions

Metadata

Runner