Crate coupe_tools

Modules

ittapi

Wrapper around ittapi to avoid spreading #cfg flags in the codebase.

Structs

EdgeWeightDistError

MeshFormatError

Problem

Enums

EdgeWeightDistribution

MeshFormat

Traits

ToRunner

Functions

barycentres

dual

The adjacency matrix that models the dual graph of the given mesh.

parse_algorithm

parse_args

Helper to handle help and version options automatically.

read_mesh

Helper to read a mesh either from stdin or from a file.

set_edge_weights

used_element_count

The number of elements that are taken into account for partitioning.

write_mesh

Helper to write a mesh, either to stdout or to a file, in the given format.

writer

Helper function to retrieve the Write implementation matching the given command-line argument.

Type Aliases

Metadata

Runner